refactor: client => frontend
