+++ website/prorata/frontend/is-promise.ts
